requirements.txt
Django==1.8.2
django-haystack==2.4.0
djangorestframework==3.1.2
djangorestframework-jwt==1.5.0
MySQL-python==1.2.5
PyJWT==1.3.0
pysolr==3.3.2
requests==2.7.0
